split up forc call lib.rs file to mod.rs and parser.rs
